Rename RDF solution summary to abstract
We are using `summary` for two different means and it has become confusing.

The content item's description is a hardcoded string in a `summary` method and the body is constructed from the RDF solution[:summary].

Now we will return the latter in a new key called `abstract` ie. RDF solution[:abstract]