Nasty mechanism to avoid 422s
We can get ourselves in a state where documents are not published,
but we need to `put_content` to them by `content_id`. The publishing
API only permits lookups of `content_id` by `base_path` for published
items, so we can leave ourselves in a state where we can't discard a
draft that is occupying a `base_path` under a different `content_id`.

If we haven't published for whatever reason, make sure we discard the
 draft.