Make Load::Outputs use AttachmentIndex
This means that if the AttachmentIndex is warmed and assets are
already in asset-manager, we won't re-download the file, we'll just
refer to what's in the index.

It's a bit hacky; the Attachment class is already overloaded with
responsibilities and we've just added to its arsenal the ability to
pretend an asset manager response came from there and not the index.