Add an OutputSerializer
To get our solutions across the wire to Sidekiq and preserve the
behaviour we want, we need to jury-rig a serialization of an
RDF::Query::Solution back to a JSON binding. This isn't available
from the RDF gems on their own but since JSON query result bindings
aren't hugely complex and we get a boost from rdf-json's extension
methods, we can do this.

The deserialization is just SPARQL::Query#parse_json_bindings